etl/python/requirements.txt
alembic==0.8.10
amqp==2.2.2
apache-airflow==1.8.2
asn1crypto==0.23.0
Babel==2.5.1
bcrypt==3.1.3
billiard==3.5.0.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.21
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.7
croniter==0.3.19
cryptography==2.0.3
dill==0.2.7.1
docutils==0.14
filechunkio==1.8
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
flower==0.9.2
funcsigs==1.0.0
future==0.15.2
gitdb2==2.0.2
GitPython==2.1.5
gunicorn==19.3.0
idna==2.6
itsdangerous==0.24
JayDeBeApi==1.1.1
Jinja2==2.8.1
jmespath==0.9.3
JPype1==0.6.2
kombu==4.1.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
numpy==1.13.1
ordereddict==1.1
pandas==0.20.3
psutil==4.4.2
psycopg2==2.7.3.1
pycparser==2.18
Pygments==2.2.0
python-daemon==2.1.2
python-dateutil==2.6.1
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.11
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.1.14
tabulate==0.7.7
thrift==0.9.3
tornado==4.5.2
Unidecode==0.4.21
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
WTForms==2.1
zope.deprecation==4.3.0
